# [SPARK-30768][SQL] Constraints inferred from inequality attributes