--- a/sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/optimizer/InferFiltersFromConstraintsSuite.scala
+++ b/sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/optimizer/InferFiltersFromConstraintsSuite.scala
@@ -402,7 +402,7 @@ class InferFiltersFromConstraintsSuite extends PlanTest {
   test("Constraints inferred from inequality attributes: case2") {
     val original = testRelation.where('a < 'b && 'b < 'c && 'c < 5)
     val optimized = testRelation.where(IsNotNull('a) && IsNotNull('b) && IsNotNull('c)
-      && 'a < 'b && 'b < 'c && 'a < 5 && 'b < 5 && 'c < 5)
+      && 'a < 'b && 'b < 'c && 'c > 'a && 'a < 5 && 'b < 5 && 'c < 5)
     comparePlans(Optimize.execute(original.analyze), optimized.analyze)
   }