### What changes were proposed in this pull request?

In our production environment, there are many queries similar to this pattern:
```sql
CREATE TABLE spark_30768_1(user_id DECIMAL(18,0), src_cre_dt DATE /*other columns*/);
CREATE TABLE spark_30768_2(user_id DECIMAL(18,0), start_dt DATE, end_dt DATE /*other columns*/);

SELECT spark_30768_1.src_cre_dt /*other columns*/                                                
FROM                                                                            
  spark_30768_1                                                                 
  INNER JOIN spark_30768_2 ON spark_30768_1.user_id = spark_30768_2.user_id     
  AND spark_30768_1.src_cre_dt >= spark_30768_2.start_dt                        
  AND spark_30768_1.src_cre_dt < spark_30768_2.end_dt                           
WHERE spark_30768_1.src_cre_dt BETWEEN date '2020-02-01' AND date '2020-02-07'      
```
In this case, we can infer more constraints to improve query performance. E.g. `spark_30768_2.start_dt <= '2020-02-07'` inferred from `spark_30768_2.start_dt <= spark_30768_1.src_cre_dt <= '2020-02-07'` and `spark_30768_2.end_dt > '2020-02-01'` inferred from `spark_30768_2.end_dt > spark_30768_1.src_cre_dt >= '2020-02-01'`. This PR add support infer these constraints from inequality attributes.

### Why are the changes needed?
Improve query performance.  Teradata support this optimization:
https://docs.teradata.com/reader/Ws7YT1jvRK2vEr1LpVURug/V~FCwD9BL7gY4ac3WwHInw?section=xcg1472241575102__application_of_transitive_closure_section


### Does this PR introduce any user-facing change?
No.


### How was this patch tested?
Unit test and benchmark test.

Benchmark code and benchmark result:
```scala
import org.apache.spark.benchmark.Benchmark
import org.apache.spark.sql.SaveMode.Overwrite

val numRows = 1024 * 1024
spark.range(numRows).selectExpr("id as c1", "id as c2").write.saveAsTable("t1")
spark.range(numRows).selectExpr("id as c1", "id as c2").write.saveAsTable("t2")

val title = "Constraints inferred from inequality constraint"
val benchmark = new Benchmark(title, numRows, minNumIters = 5)
benchmark.addCase(s"t1.c1 > ${numRows - 1000}") { _ =>
  spark.sql(s"select count(*) from t1 join t2 on (t1.c1 > t2.c1 and t2.c1 > ${numRows - 1000})").write.format("noop").mode(Overwrite).save()
}

benchmark.addCase("t1.c1 < 1000") { _ =>
  spark.sql("select count(*) from t1 join t2 on (t1.c1 < t2.c1 and t2.c1 < 1000)").write.format("noop").mode(Overwrite).save()
}
benchmark.run()
```
Before this PR:
```
Java HotSpot(TM) 64-Bit Server VM 1.8.0_191-b12 on Mac OS X 10.13.6
Intel(R) Core(TM) i7-7820HQ CPU @ 2.90GHz
Constraints inferred from inequality constraint:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
t1.c1 > 1047576                                   30519          31171         567          0.0       29105.2       1.0X
t1.c1 < 1000                                      50525          50906         331          0.0       48184.7       0.6X
```

After this PR:
```
Java HotSpot(TM) 64-Bit Server VM 1.8.0_191-b12 on Mac OS X 10.13.6
Intel(R) Core(TM) i7-7820HQ CPU @ 2.90GHz
Constraints inferred from inequality constraint:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
t1.c1 > 1047576                                     179            191          15          5.9         170.5       1.0X
t1.c1 < 1000                                        175            195          27          6.0         166.7       1.0X
```

Also test this feature in our production environment, it can significantly improve the query performance of at least 6 SQLs (a total of 200 SQLs):

SQL ID | Before this PR(seconds) | After this PR(seconds)
-- | -- | --
323 | 108 | 73
368 | 56 | 26
372 | 72 | 28
373 | 60 | 26
374 | 113 | 44
375 | 82 | 55

For e.g. SQL 372. It prevents (18,413,424,580 - 162,205,133 = 18,251,219,447) rows from participating in shuffle:
Before this PR:
![before](https://user-images.githubusercontent.com/5399861/75850979-8404e100-5e23-11ea-8ea0-9fd4347ac200.png)

After this PR:
![after](https://user-images.githubusercontent.com/5399861/75851328-5704fe00-5e24-11ea-9015-41c63284fafa.png)