@wangyum 2020-02-12
@peter-toth https://github.com/apache/spark/pull/27518/commits/999126cf899816e4c4f2035f70c9099fceb63516 helps the example query.