@peter-toth 2020-02-13
> @peter-toth [999126c](https://github.com/apache/spark/commit/999126cf899816e4c4f2035f70c9099fceb63516) helps the example query.

Thanks @wangyum. I see now that you basically reverted your first commit.